ForgedApp-Backend/requirements.txt
absl-py==0.11.0
astor==0.8.1
boto3==1.17.7
botocore==1.20.7
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dataclasses==0.8
fastapi==0.63.0
Flask==1.1.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.35.0
h11==0.12.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.0.0.21
Pillow==8.1.0
protobuf==3.14.0
pydantic==1.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.4
scipy==1.5.4
six==1.15.0
starlette==0.13.6
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
